requirements.txt
faiss-cpu
sentence-transformers
plotly
rapidfuzz
nltk
networkx
pyvis
chardet
openpyxl
databricks-sql-connector